vm/discord-bot/requirements.txt
discord.py[voice]==2.4.0
aiohttp==3.10.5
httpx==0.27.0
python-dotenv==1.0.1
